refactor(podman): grouping remote files
