+++ extensions/podman/packages/extension/src/remote/podman-remote-connections.ts
@@ -22,7 +22,7 @@ import { join } from 'node:path';
 
 import * as extensionApi from '@podman-desktop/api';
 
-import { getPodmanCli } from './podman-cli';
+import { getPodmanCli } from '../podman-cli';
 import { PodmanRemoteSshTunnel } from './podman-remote-ssh-tunnel';
 
 interface ConnectionListFormatJson {
